backend/pyproject.toml
[project]
name = "speech-assist-backend"
version = "0.1.0"
description = "Backend for the Speech Ability Assistance System."
requires-python = "==3.10.18"

dependencies = [
   "fastapi[all]",
   "opencv-python",
   "mediapipe",
   "paho-mqtt",
   "httpx",
   "realtimestt",
   "openai>=1.107.0",
]


